requirements.txt
async-timeout==5.0.1
blinker==1.8.2
click==8.1.8
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
pillow==10.4.0
redis==5.2.1
Werkzeug==3.0.6
zipp==3.20.2
